refactor: 프로필 북마크게시물 파일구조 수정 (#402)
